Updates imports and paths after package move
