+++ internal/controllers/suite_test.go
@@ -54,8 +54,8 @@ var _ = BeforeSuite(func() {
 	By("bootstrapping test environment")
 	testEnv = &envtest.Environment{
 		CRDDirectoryPaths: []string{
-			filepath.Join("..", "config", "crd", "bases"),
-			filepath.Join("..", "testdata", "crds")},
+			filepath.Join("..", "..", "config", "crd", "bases"),
+			filepath.Join("..", "..", "testdata", "crds")},
 		ErrorIfCRDPathMissing: true,
 	}
